Ditching iframe, back to eval solution
In this way scripts included in the main code would just work. This
would help the 'sandbox' context.
It might be worth later to try with iframe, with code to include the
libraries needed for the code execution so everything is separated by
the main application.

Might be difficult to handle tfjs-vis like this:
https://storage.googleapis.com/tfjs-vis/mnist/dist/index.html